project_setup_and_basic_token_contract/tests.rs
```rust
// Import the necessary modules
use anchor_lang::prelude::*;
use anchor_spl::token::{self, TokenAccount, Transfer};
use project_setup_and_basic_token_contract::TokenContract;

// Define the test context
#[derive(Accounts)]
pub struct TestContext<'info> {
    #[account(signer)]
    pub user: AccountInfo<'info>,
    #[account(mut)]
    pub from: AccountInfo<'info>,
    #[account(mut)]
    pub to: AccountInfo<'info>,
    #[account(mut)]
    pub token_program: AccountInfo<'info>,
}

// Define the test module
#[mod_tests]
mod tests {
    use super::*;

    // Test the mintToken function
    #[tokio::test]
    async fn test_mint_token() {
        let ctx = TestContext::new().await;
        let cpi_accounts = Transfer {
            from: ctx.accounts.from.clone(),
            to: ctx.accounts.to.clone(),
            authority: ctx.accounts.user.clone(),
        };
        let amount = 100;
        let cpi_program = ctx.accounts.token_program.clone();
        let cpi_ctx = CpiContext::new(cpi_program, cpi_accounts);
        token::transfer(cpi_ctx, amount).unwrap();
        assert_eq!(TokenContract::mintToken(ctx.accounts.into(), amount).unwrap(), ());
    }

    // Test the transferToken function
    #[tokio::test]
    async fn test_transfer_token() {
        let ctx = TestContext::new().await;
        let cpi_accounts = Transfer {
            from: ctx.accounts.from.clone(),
            to: ctx.accounts.to.clone(),
            authority: ctx.accounts.user.clone(),
        };
        let amount = 50;
        let cpi_program = ctx.accounts.token_program.clone();
        let cpi_ctx = CpiContext::new(cpi_program, cpi_accounts);
        token::transfer(cpi_ctx, amount).unwrap();
        assert_eq!(TokenContract::transferToken(ctx.accounts.into(), amount).unwrap(), ());
    }

    // Test the burnToken function
    #[tokio::test]
    async fn test_burn_token() {
        let ctx = TestContext::new().await;
        let cpi_accounts = Transfer {
            from: ctx.accounts.from.clone(),
            to: ctx.accounts.to.clone(),
            authority: ctx.accounts.user.clone(),
        };
        let amount = 50;
        let cpi_program = ctx.accounts.token_program.clone();
        let cpi_ctx = CpiContext::new(cpi_program, cpi_accounts);
        token::transfer(cpi_ctx, amount).unwrap();
        assert_eq!(TokenContract::burnToken(ctx.accounts.into(), amount).unwrap(), ());
    }
}
```